moved everything to one globals.css file
