+++ live-site/package.json
@@ -18,6 +18,5 @@
   "dependencies": {
     "gatsby": "^4.20.0",
     "gatsby-plugin-mailchimp": "^5.2.2"
-
   }
 }
