+++ shared-ui/components/404/404.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import PrimaryButton from '../../../shared-ui/components/primary-button/PrimaryButton';
 import { min } from '../../../shared-ui/lib/responsive';
 import useMatchMedia from 'react-use-match-media';
-import '../../globals.css';
+import '../../style/globals.css';
 import { StyledH1, StyledH4, ButtonContainer } from './404.styles';
 
 const NotFound: React.FC = () => {
