+++ shared-ui/components/meet-the-team/mobile-team-column/MobileTeamColumn.tsx
@@ -1,6 +1,10 @@
 import React, { useState } from 'react';
 import Arrow from '../../arrow/Arrow';
-import { MobileTeamColumnProps, Person, TeamColumnInfo } from '../../../lib/types';
+import {
+  MobileTeamColumnProps,
+  Person,
+  TeamColumnInfo
+} from '../../../lib/types';
 import { getLeftOrRight } from '../../../../main-site/lib/utils';
 import {
   StyledMobileTeamColumn,
