--- shared-ui/index.css
@@ -1,29 +1,0 @@
-@font-face {
-  font-family: 'Calistoga-Regular';
-  src: local('Calistoga-Regular'),
-    url('./style/fonts/Calistoga-Regular.ttf') format('truetype');
-}
-
-@font-face {
-  font-family: 'NunitoSans-Bold';
-  src: local('NunitoSans-Bold'),
-    url('./style/fonts/NunitoSans-Bold.ttf') format('truetype');
-}
-
-@font-face {
-  font-family: 'NunitoSans-Light';
-  src: local('NunitoSans-Light'),
-    url('./style/fonts/NunitoSans-Light.ttf') format('truetype');
-}
-
-@font-face {
-  font-family: 'NunitoSans-Regular';
-  src: local('NunitoSans-Regular'),
-    url('./style/fonts/NunitoSans-Regular.ttf') format('truetype');
-}
-
-@font-face {
-  font-family: 'NunitoSans-SemiBold';
-  src: local('NunitoSans-SemiBold'),
-    url('./style/fonts/NunitoSans-SemiBold.ttf') format('truetype');
-}
