+++ shared-ui/style/colors.ts
@@ -21,13 +21,11 @@ const colors = {
   NIGHT_HEADER_COLOR: '#628BC6',
   TEXT_BROWN: '#634C1B',
   TEXT_GREY: '#D9D9D9',
-  DIRECTORS_YELLOW: '#F2C94C', 
-  TECH_BLUE: '#2F80ED', 
-  DESIGN_PINK: '#BB6BD9', 
-  SOORANGE: '#F2A32B', 
+  DIRECTORS_YELLOW: '#F2C94C',
+  TECH_BLUE: '#2F80ED',
+  DESIGN_PINK: '#BB6BD9',
+  SOORANGE: '#F2A32B',
   SPONSIBGREEN: '#27AE60'
-
-
 };
 
 export { colors };
