--- shared-ui/style/global.css
@@ -1,8 +1,0 @@
-body {
-  margin: 0;
-  max-width: 1920px;
-}
-
-html {
-  max-width: 1920px;
-}
