+++ shared-ui/style/globals.css
@@ -1,0 +1,67 @@
+.day-background-desktop {
+  background-image: url(../images/day-background-desktop.svg);
+  background-size: cover;
+}
+
+.dark-background-desktop {
+  background-image: url(../images/dark-background-desktop.svg);
+  background-size: cover;
+}
+
+.mobile-light-background {
+  background-image: url(../images/mobile-light-background.svg);
+  background-size: cover;
+}
+
+.mobile-dark-background {
+  background-image: url(../images/mobile-dark-background.svg);
+  background-size: cover;
+}
+
+.error-background-desktop {
+  background-image: url(../images/error-desktop-background.svg);
+  background-size: cover;
+  height: 100vh;
+}
+
+.error-background-mobile {
+  background-image: url(../images/error-mobile-background.svg);
+  background-size: cover;
+  height: 100vh;
+}
+
+html,
+body {
+  margin: 0;
+  max-width: 1920px;
+}
+
+@font-face {
+  font-family: 'Calistoga-Regular';
+  src: local('Calistoga-Regular'),
+    url('./fonts/Calistoga-Regular.ttf') format('truetype');
+}
+
+@font-face {
+  font-family: 'NunitoSans-Bold';
+  src: local('NunitoSans-Bold'),
+    url('./fonts/NunitoSans-Bold.ttf') format('truetype');
+}
+
+@font-face {
+  font-family: 'NunitoSans-Light';
+  src: local('NunitoSans-Light'),
+    url('./fonts/NunitoSans-Light.ttf') format('truetype');
+}
+
+@font-face {
+  font-family: 'NunitoSans-Regular';
+  src: local('NunitoSans-Regular'),
+    url('./fonts/NunitoSans-Regular.ttf') format('truetype');
+}
+
+@font-face {
+  font-family: 'NunitoSans-SemiBold';
+  src: local('NunitoSans-SemiBold'),
+    url('./fonts/NunitoSans-SemiBold.ttf') format('truetype');
+}
